Update README.md
Planned updates signify a comprehensive effort to bring Trinity Graph Engine to the forefront of semantic, ontological, and distributed computing. The full-sample programs will serve as invaluable resources for developers, especially those navigating the complexities of graph-based data structures, distributed systems, and ontological frameworks. The new enhancements promise to make TGE a cutting-edge tool for a wide range of applications.

Signed-off-by: Tavi Truman <TaviTruman@users.noreply.github.com>